functionality to add new types and floor selection menu
When page is loaded the user is first
prompted with a floor selection. The user must click edit to load up page. Also now when a type is added it will now show up in the drop down menu. additionally on the drop down it now lets the user know to press enter to create a new type/floor